fix: バリデーションとエラーメッセージに関する修正
