+++ backend/model/error_res.go
@@ -1,0 +1,6 @@
+package model
+
+type ErrorResponse struct{
+	Result  bool   `json:"result"`
+	Message string `json:"message"`
+}
